client: release: deploy with new feedstock crud
