+++ client/src/utils/translate/shared-translate.ts
@@ -1,27 +1,36 @@
-import { MeasureUnit } from "@/types/measure/measure-unit"
+import { MeasureUnit, StateMatter } from "@/types/measure/measure-unit";
 
-// Traducciones de unidades
-const translationMeasureUnit: Record<MeasureUnit, string> = {
+
+
+// Traducciones para estados de materia
+const translationsStateMatter: Record<StateMatter, string> = {
+  [StateMatter.SOLID]: "Sólido",
+  [StateMatter.LIQUID]: "Líquido",
+  [StateMatter.GAS]: "Gas",
+};
+
+// Traducciones para unidades de medida
+const translationsMeasureUnit: Record<MeasureUnit, string> = {
   // Sólidos
-  GRAMS: "Gramos",
-  KILOGRAMS: "Kilogramos",
-  TONNES: "Toneladas",
-  UNITS: "Unidades",
-  BOXES: "Cajas",
-  METERS: "Metros",
-  SQUARE_METERS: "Metros cuadrados",
+  [MeasureUnit.GRAMS]: "Gramos",
+  [MeasureUnit.KILOGRAMS]: "Kilogramos",
+  [MeasureUnit.TONNES]: "Toneladas",
+  [MeasureUnit.UNITS]: "Unidades",
+  [MeasureUnit.BOXES]: "Cajas",
+  [MeasureUnit.METERS]: "Metros",
+  [MeasureUnit.SQUARE_METERS]: "Metros cuadrados",
 
   // Líquidos
-  LITERS: "Litros",
-  MILLILITERS: "Mililitros",
-  GALLONS: "Galones",
+  [MeasureUnit.LITERS]: "Litros",
+  [MeasureUnit.MILLILITERS]: "Mililitros",
+  [MeasureUnit.GALLONS]: "Galones",
 
   // Gases
-  CUBIC_METERS: "Metros cúbicos",
-  LITERS_GAS: "Litros (gas)",
-}
+  [MeasureUnit.CUBIC_METERS]: "Metros cúbicos",
+  [MeasureUnit.LITERS_GAS]: "Litros (gas)",
+};
+
+export const translateStateMatter = (state: StateMatter) => translationsStateMatter[state];
+export const translateMeasureUnit = (unit: MeasureUnit) => translationsMeasureUnit[unit];
 
 
-export const translateMeasureUnit = (mu: MeasureUnit) => {
-  return translationMeasureUnit[mu]
-}
